add content data feed and security privacy
