+++ src/pages/terms-and-conditions/_copyright.js
@@ -1,0 +1,37 @@
+import React from 'react'
+import { Header } from 'components/elements/typography'
+import { StyledSection, TncText } from './_styles'
+import { localize } from 'components/localization'
+
+export const Copyright = () => (
+    <StyledSection>
+        <Header as="h2" color="black">
+            {localize('Copyright')}
+        </Header>
+        <TncText>
+            {localize(
+                'The Deriv.com website is protected by certain copyrights.',
+            )}
+        </TncText>
+        <TncText>
+            {localize(
+                'The materials comprising the Deriv.com website (including without limitation all articles, text, images, logos, compilation, systems, code, and design) are under Copyright 1999-2019 Binary Group of companies with all rights reserved.',
+            )}
+        </TncText>
+        <TncText>
+            {localize(
+                'Such materials may be copied and distributed on a limited basis for noncommercial purposes only, provided that any material copied remains intact and that all copies include the following notice in a clearly visible position: “Copyright 1999-2019 Binary Ltd. All rights reserved.”',
+            )}
+        </TncText>
+        <TncText>
+            {localize(
+                'These materials may not be copied or redistributed for commercial purposes or for compensation of any kind without prior written permission from a Binary Group company.',
+            )}
+        </TncText>
+        <TncText>
+            {localize(
+                'Deriv.com and the bull/bear logo are registered trademarks.',
+            )}
+        </TncText>
+    </StyledSection>
+)
