+++ src/pages/terms-and-conditions/_styles.js
@@ -1,0 +1,28 @@
+import styled from 'styled-components'
+import { Header, Text } from 'components/elements/typography'
+
+export const StyledSection = styled.section`
+    padding-left: 8rem;
+`
+
+export const TncText = styled(Text)`
+    margin: 2rem 0;
+`
+
+export const TncListText = styled(Text)`
+    margin: 0.8rem 0;
+`
+
+export const StyledHeader = styled(Header)`
+    margin-top: 3.2rem;
+`
+
+export const StyledOrder = styled.ol`
+    list-style: ${props => props.type || ''};
+    font-size: var(--text-size-s);
+    padding-left: 1.8rem;
+
+    & > li {
+        padding-left: 0.5rem;
+    }
+`
